requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair==4.0
pinecone-client==2.2.2
tiktoken
OpenAI
pypdf
flask_cors
flask
streamlit-lottie
Pillow



